requirements.txt
anthropic>=0.40.0
openai>=1.54.0
google-genai>=0.2.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
Pillow>=10.0.0
